add TempleteService delete Template realisation classes

+++ militarism2/src/main/java/com/militarism2/militarism2/game/unit/service/TemplateService.java
@@ -6,5 +6,5 @@ import com.militarism2.militarism2.game.unit.army.template.Template;
 
 
 public interface TemplateService {
-	Optional<Template> getUserById(int id);
+	Optional<Template> getUserById(long id);
 }
